Add files via upload
SQL, SQLite, GoogleCollab:
Link: https://colab.research.google.com/drive/111k5VqfIEbk4of4P5g6RCVCk1CjnFwVe?usp=sharing


Meanwhile, I have been joining the head of course content in some feedback interviews with learners. The positive interviews always seem to have two flavors: a few learners preferred courses that ran longer and dug deeper into an area, while others preferred courses that were quick and exposed them to a new area. I decide to investigate this and see if somehow the nps was correlated to this combination of the length of the course and the level of the course.